# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: netmoded — HTTP API панели управления роутером
  version: "1.0.0-phase1"
  description: |
    Единственный источник правды для обоих треков разработки: Go-обработчики и
    фронтенд пишутся по этому файлу, а не друг по другу.

    ## Что нужно знать до чтения

    - **Только LAN.** Сервер слушает адрес из `netmode.main.listen`
      (сид `192.168.9.1`). `0.0.0.0` запрещён. ADR-0014.
    - **Токен** `Authorization: Bearer <…>` обязателен на всех `/api/*`.
      Файл `/etc/netmoded/token`, права 0600.
    - **`/api/status` опрашивается 1 Гц** (SPEC §7). Внешние вызовы внутри него
      кэшируются 500 мс. WebSocket и SSE не используются — ADR-0017.
    - **Формат ошибок один на все коды** — см. `docs/contracts/errors.md`
      и схему `Error`.
    - **Пароли только на запись.** В ответах есть `has_key: bool`, поля `key`
      нет никогда. ADR-0012.
    - **Идентификатор сети — имя секции UCI** (`wifinet0`), не индекс
      `@wifi-iface[N]`. ADR-0005.
    - **Фаза 1 не меняет поведение станционного радио.** Демон трогает только
      выключенные секции и не вызывает применения. ADR-0009. Какое радио
      станционное — выводится из системы, а не задано константой; не
      выяснилось — `503 radio_unknown`. ADR-0019.
    - Времена — RFC 3339 в UTC (`…Z`).

    ## Эндпоинты SPEC §7, которых здесь нет

    `POST /api/confirm` — **удалён**. Он подтверждал взведённый откат;
    автоматического отката нет ни в какой форме (ADR-0006), подтверждать
...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /api/status:
    get:
      tags: [status]
      summary: Полная картина состояния
      description: |
        Опрашивается фронтендом раз в секунду. **Никогда не отвечает 409, 501
        или 503**: недоступность внешних компонентов выражается полями
        (`b4.available`, `nikki.available`, `associated_ssid: null`), а не
        кодом. Иначе лежащий b4 ослеплял бы всю панель.

        Внешние вызовы (ubus, b4, Clash API) кэшируются 500 мс; `uci show
        wireless` читается каждый раз — из него берётся `wireless_fingerprint`.
      responses:
        "200":
          description: Состояние
          headers:
            Cache-Control:
              schema: { type: string, const: "no-store" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
              examples:
                single:       { $ref: "#/components/examples/StatusSingle" }
                ambiguous:    { $ref: "#/components/examples/StatusAmbiguous" }
                allDisabled:  { $ref: "#/components/examples/StatusAllDisabled" }
                empty:        { $ref: "#/components/examples/StatusEmpty" }
                b4Down:       { $ref: "#/components/examples/StatusB4Unavailable" }
                jobRunning:   { $ref: "#/components/examples/StatusJobRunning" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/mode:
    post:
      tags: [mode]
      summary: Сменить режим обработки трафика
      description: |
        Медленная операция (5–15 с): `uci set netmode.main.mode` + commit,
        затем `/usr/local/bin/netmode-apply <mode>`. Джоб и светодиод —
        ADR-0013, SPEC §8.

        UCI пишется **до** запуска скрипта: конфиг хранит намерение, скрипт
        приводит систему к намерению. Если скрипт упал, намерение сохранено,
        джоб в состоянии `failed`, повтор возможен. Отката нет — ADR-0006.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              additionalProperties: false
              properties:
                mode:
                  type: string
                  enum: [nikki, b4, off]
      responses:
        "202":
          description: Джоб принят
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: |
            `job_busy` — уже выполняется джоб. Очереди нет сознательно
            (SPEC §6): второй запрос отбивается.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/upstream:
    post:
      tags: [wifi]
      summary: Сменить активную upstream-сеть — ОТЛОЖЕНО ДО ФАЗЫ 2
      description: |
        **Всегда отвечает 501 в этой поставке.** Форма запроса заморожена
        сейчас, чтобы фронтенд не переписывался в фазе 2.

        Причины отсрочки (ADR-0016): нет автоматического отката (ADR-0006);
        не отвечены RQ-03 (существует ли применение только к radio0, или
        домашняя точка `grenderNet` моргает — радио делят одну phy0) и RQ-04
        (рвёт ли скан активную ассоциацию).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              additionalProperties: false
              properties:
                id:
                  type: string
                  description: Имя секции UCI сохранённой сети
                  examples: ["wifinet2"]
      responses:
        "501":
          description: Не реализовано в этой поставке
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                code: not_implemented
                error: >-
                  Переключение внешней сети появится во второй фазе: сначала
                  надо выяснить, можно ли применить изменение только к
                  станционному радио, не уронив домашнюю сеть — оба радио делят
                  одну phy0 (RQ-03).
  /api/wifi/scan:
    get:
      tags: [wifi]
      summary: Скан эфира
      description: |
        `ubus call iwinfo scan` по интерфейсу станции. И станционное радио, и
        имя интерфейса **выводятся** из `network.wireless status`
        (`<radio>.interfaces[]` — `config.mode == "sta"`, оттуда же `ifname` и
        `config.band`), не хардкодятся — `scripts/check-no-hardcoded-if.sh`,
        порядок разрешения в ADR-0019.

        Станция работает в ОДНОМ диапазоне, поэтому **скан физически не может
        показать сети другого**. Поле `band` и текст `note` существуют, чтобы
        панель это написала: отсутствие знакомой сети иначе выглядит как баг.
        На снимке роутера станция на 2.4 ГГц, но это факт снимка, а не
        константа: `band` отдаёт то, что выяснено, а `note` следует за ним.

        Синхронный, без джоба, хотя и медленный: ничего не меняет, результат
        нужен целиком (ADR-0013).

        Фикстура ответа снята — `docs/recon/raw/23-ubus-iwinfo-scan.json`,
        разбор проверяется на ней в `wireless.ParseScan`. RQ-04 (рвёт ли скан
        активную ассоциацию) закрыт **косвенно**: скан выполнялся при живой
        ассоциации, и она пережила его. Прямой проверки не было.
      responses:
        "200":
          description: Результат скана
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503":
          description: |
            `radio_unknown` — не выяснено, какое радио работает станцией: ни
            в `network.wireless status`, ни в `/etc/config/wireless` нет
            интерфейса с `mode=sta`. Демон не берёт радио наугад (ADR-0019).

            `ifname_unknown` — радио выяснено, но имя его интерфейса не
            выводится (радио выключено, ubus не ответил).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/wifi/networks:
    get:
      tags: [wifi]
      summary: Сохранённые upstream-сети
      description: |
        Секции `wifi-iface` из `/etc/config/wireless` с `mode=sta` на
        станционном радио (ADR-0003). Какое радио станционное — **выводится**
        из системы, а не задано: `mode=sta` в `network.wireless status`, при
        выключенном радио — секция `wifi-iface` в UCI (ADR-0019). Домашняя
        точка доступа (на снимке роутера — `wifinet1`, `grenderNet`,
        `mode=ap`) в список не входит на любом радио.

        Порядок — как в файле. `ETag` и поле `fingerprint` содержат одно и то
        же значение; оно требуется в `If-Match` при записи.
      responses:
        "200":
          description: Список
          headers:
            ETag:
              schema: { type: string }
              description: Отпечаток `/etc/config/wireless`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NetworksResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503":
          description: |
            `radio_unknown` — не выяснено, какое радио работает станцией.
            Демон не берёт радио наугад (ADR-0019).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      tags: [wifi]
      summary: Создать или изменить сохранённую сеть
      description: |
        Быстрая операция: `uci set` + `uci commit wireless`, десятки
        миллисекунд, синхронный ответ. **Применения нет** — ни `wifi reload`,
        ни джоба, ни светодиода (ADR-0009).

        Без `id` — создание. Секция создаётся именованной (`netmode_<8 hex>`)
        и **всегда с `disabled='1'`**: отсутствие `disabled` означает
        «включена», а создать включённую станционную секцию фаза 1 не вправе.
        Это единственное место, где демон вообще пишет `disabled`.

        С `id` — правка существующей. Разрешена только для **выключенной**
        секции; правка включённой отбивается `409 enabled_network_readonly`.

        `key` отсутствует в теле → пароль не меняется. `key: ""` → записывается
        пустой пароль (осмысленно при `encryption: "none"`). Это разные вещи.

        **Незнакомые поля отвергаются** — `400 unsupported_field`. Демон не
        принимает того, чего не понимает: молча пропущенное поле выглядит как
        выполненная просьба. В частности `network` (выбор L3-сети) — это
        смена upstream, отложенная до фазы 2 (ADR-0016).
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NetworkWrite" }
      responses:
        "200":
          description: Записано. Тело — обновлённый список, как у `GET`.
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NetworksResponse" }
        "400":
          description: |
            `bad_request` — тело не разбирается, нет обязательного поля,
            значение вне набора.
            `unsupported_field` — в теле поле, которого нет в `NetworkWrite`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/WifiWriteConflict" }
        "503":
          description: |
            `radio_unknown` — не выяснено, какое радио работает станцией.
            Демон не берёт радио наугад (ADR-0019).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/wifi/networks/{id}:
    delete:
      tags: [wifi]
      summary: Удалить сохранённую сеть
      description: |
        Разрешено только для выключенной секции. Удаление включённой —
        `409 enabled_network_readonly` (ADR-0009). Как и правка: `uci delete` +
        `uci commit`, без применения.

        `id` в пути — имя секции. Индексы `@wifi-iface[N]` не принимаются
        никогда: они сдвигаются при удалении в LuCI и указали бы на чужую сеть
        (ADR-0005).
      parameters:
        - $ref: "#/components/parameters/NetworkId"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Удалено. Тело — обновлённый список.
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NetworksResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/WifiWriteConflict" }
        "503":
          description: |
            `radio_unknown` — не выяснено, какое радио работает станцией.
            Демон не берёт радио наугад (ADR-0019).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/nikki/proxies:
    get:
      tags: [nikki]
      summary: Узлы группы PROXY с задержками
      description: |
        **RQ-01 закрыт**, эндпоинт реализован и ходит в живой Clash API
        (порт 9090, секрет из `nikki.mixin.api_secret`, `Authorization:
        Bearer`). `503 nikki_unavailable` означает ровно то, что написано:
        Clash API не ответил — а не «не реализовано».

        `503 group_missing` — Clash API ответил, но группы `PROXY` в профиле
        нет. Отдельный код, потому что чинится это правкой профиля, а не
        ожиданием.

        По SPEC §7 группа `PROXY` должна быть `selector`, а не `url-test`, и
        обязана содержать пункт `AUTO` (вложенный `url-test`) — он же
        состояние по умолчанию. Без `AUTO` ломается fail-closed: группа
        `BYPASS` видит `PROXY` как один элемент и не отличает «выбран мёртвый
        узел» от «всё умерло».
      responses:
        "200":
          description: Узлы
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProxiesResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/NikkiError" }
        "503": { $ref: "#/components/responses/NikkiUnavailable" }
...
  /api/nikki/proxy:
    post:
      tags: [nikki]
      summary: Выбрать узел (быстрая операция)
      description: |
        Быстрая: Clash API, ~0.2 с, без джоба и без светодиода (SPEC §5).

        Ручной выбор трактуется как временное действие. Возврат к
        автоматическому — тот же запрос с `{"name": "AUTO"}`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              additionalProperties: false
              properties:
                name:
                  type: string
                  description: Имя узла или `AUTO` для возврата к url-test
                  examples: ["AUTO"]
      responses:
        "200":
          description: Выбрано. Тело — как у `GET /api/nikki/proxies`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProxiesResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            `group_not_selectable` — группа не принимает ручной выбор. Это
            `409`, а не `501`: конфликт с текущим профилем mihomo, а не
            отсутствующая в демоне возможность. Чинится правкой профиля.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502": { $ref: "#/components/responses/NikkiError" }
        "503": { $ref: "#/components/responses/NikkiUnavailable" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Токен из `/etc/netmoded/token` (0600). Сравнение постоянного времени.
        Заголовки CORS не отдаются никогда.

        Токен принимается тремя способами (`Server.tokenOK`): заголовком
        `Authorization: Bearer`, параметром `?token=` в адресе и кукой
        `netmode_token`. Кука ставится сервером, когда токен пришёл ссылкой:
        за `<script src>` браузер не отправляет ни заголовок, ни query, и без
        куки панель грузилась бы с `200` на главной и `401` на всех запросах.
        Статика тоже под токеном — она раскрывает имена сетей.

        **Проверки `Origin` нет.** ADR-0014 описывает `403 forbidden_origin`
        как существующий второй механизм защиты от CSRF; в коде его нет
        (`grep Origin internal/` пуст). Единственная защита — `SameSite=Strict`
        у куки. Это известное и не закрытое расхождение.
...